feat(#15, #20): add UI and filters for the beers
TODO:
- Add support for sorting the beers by property
