+++ src/pages/supabaseAuth.tsx
@@ -1,0 +1,23 @@
+import {useState, useEffect} from "react";
+
+import {supabase} from "utils/supabaseClient";
+import Auth from "src/components/Auth";
+import Account from "src/components/Account";
+
+export default function Home() {
+  const [session, setSession] = useState(null);
+
+  useEffect(() => {
+    setSession(supabase.auth.session());
+
+    supabase.auth.onAuthStateChange((_event, session) => {
+      setSession(session);
+    });
+  }, []);
+
+  return (
+    <div className="container" style={{padding: "50px 0 100px 0"}}>
+      {!session ? <Auth /> : <Account key={session.user.id} session={session} />}
+    </div>
+  );
+}
