+++ utils/helpers.ts
@@ -1,7 +1,8 @@
-import {IBeer, IFilterBrand} from 'src/pages/beersOld'
+import {IBeer, IFilterBrand} from 'src/pages/index'
+import { colors } from "utils/constant";
 
 /**
- * It takes the array of beers and reduces it to an array of unique values
+ * It takes the array of beers and reduces it to an array of unique values with IFIlterBrand object shape
  * It works for 
  * 
  * @param beers as IBeer[]
@@ -10,11 +11,11 @@ import {IBeer, IFilterBrand} from 'src/pages/beersOld'
  */
 export const uniqueValues = function (beers : IBeer[], valueToLookFor: keyof IBeer) : IFilterBrand[] {
   return beers.reduce((prev, curr) => {
-      if (!prev.some((element) => element.value === curr[valueToLookFor])) {
+      if (curr[valueToLookFor] && !prev.some((element) => element.value === curr[valueToLookFor])) {
         prev.push({
           value: String(curr[valueToLookFor]),
           label: String(curr[valueToLookFor]),
-          src: "",
+          src: valueToLookFor === "brand" ? String(curr.imagen) : "",
         });
       }
